requirements.txt
# Bridge Server Dependencies
# Python 3.10+

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP Client
httpx==0.25.1

# Cryptography
pycryptodome==3.19.0
python-jose[cryptography]==3.3.0

# Session Storage
redis==5.0.1
hiredis==2.2.3

# Environment Variables
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1  # For async testing

# Development
black==23.11.0
ruff==0.1.6
mypy==1.7.1
